requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers>=4.3,<5.0
psycopg2-binary==2.9.10
bcrypt==4.2.1
drf-spectacular>=0.27,<1.0
gunicorn==23.0.0
python-dotenv==1.0.1
